admin_panel

+++ app/Http/Controllers/PresentationController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Conference;
 use Illuminate\Http\Request;
 use App\Models\Presentation;
 use App\Models\User;
@@ -16,19 +17,22 @@ class PresentationController extends Controller
     }
 
     public function edit (Presentation $presentation){
+        $this->authorize('update', $presentation);
         return view('presentations/edit-p', compact('presentation'));
 
     }
 
     public function patch(Presentation $presentation){
+        $this->authorize('update', $presentation);
         $data = request()->validate([
             'date' => 'required|date',
+            'room_id' => 'required',
             'start' => 'required',
             'end' => 'required|after:start',
         ]);
         $presentation->accepted = 1;
         $presentation->update($data);
-        return redirect("/presentation/pending");
+        return redirect("conference/{$presentation->conference()->first()->id}/presentations");
         
     }
 
@@ -38,13 +42,34 @@ class PresentationController extends Controller
 
     }
 
+    public function showall(Conference $conference){
+
+        return view('presentations/showall-p', compact('conference'));
+
+    }
+
+    public function show(Presentation $presentation){
+
+        return view('presentations/show-p', compact('presentation'));
+
+    }
+
     public function accept(Presentation $presentation){
+        $this->authorize('update', $presentation);
         
         $presentation->update();
         return redirect('presentation/pending');
 
     }
+
     public function delete(Presentation $presentation){
+        $this->authorize('update', $presentation);
+        Presentation::destroy($presentation->id);
+        return redirect("conference/{$presentation->conference()->first()->id}/presentations");
+
+    }
+    public function reject(Presentation $presentation){
+        $this->authorize('update', $presentation);
         Presentation::destroy($presentation->id);
         return redirect('presentation/pending');
